Rename newResources.ts and tests to no longer say "new".
